Bug fix for landscape mode.
Im still too much of an android newb to understand exactly why this
happens... but for some reason my default project creates a layout-land
folder inside res. So when I switch to landscape mode the fragment looks
in this folder which doesn't have the same elements my code is
expecting.

My temporary fix is to just copy the fragment and other files from
layout to layout-land but theres probably a better way to get some code
reuse here.